[refactor] repository
